+++ src/by/it/ausiyevich/lesson3/TaskB1.java
@@ -26,10 +26,10 @@ float f - присваивание в десятичном формате
 strictfp class TaskB1 {
     public static  void main(String[] args) {
         double d = 111.111;
-        int b =0b111;
-        int o =0111;
+        byte b =0b111;
+        long o =0111;
         int h =0x111;
-        double f = 111.111;
+        float f = 111.111f;
         System.out.println(d + b + o + h + f);
         System.out.println(d + " " + b + " " + o + " " + h + " " + f);
     }
